feat/ restructured and grouped hooks directory
