+++ src/__tests__/hooks/useFetchCompareResults.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react';
 
-import useFetchCompareResults from '../../hooks/useFetchCompareResults';
+import useFetchCompareResults from '../../hooks/CompareResults/useFetchCompareResults';
 import getTestData from '../utils/fixtures';
 import { StoreProvider } from '../utils/setupTests';
