+++ src/components/CompareResults/CompareResultsTable.tsx
@@ -7,7 +7,7 @@ import Table from '@mui/material/Table';
 import TableContainer from '@mui/material/TableContainer';
 
 import { RootState } from '../../common/store';
-import { useAppSelector } from '../../hooks/app';
+import { useAppSelector } from '../../hooks/utils/app';
 import type { CompareResultsState } from '../../types/state';
 import CompareTableStatus from './CompareTableStatus';
 import PaginatedCompareResults from './PaginatedCompareResults';
