+++ src/components/CompareResults/FilterOptionsList.tsx
@@ -9,8 +9,8 @@ import {
 } from '@mui/material';
 
 import { RootState } from '../../common/store';
-import { useAppSelector } from '../../hooks/app';
-import useFilterCompareResults from '../../hooks/useFilterCompareResults';
+import useFilterCompareResults from '../../hooks/CompareResults/useFilterCompareResults';
+import { useAppSelector } from '../../hooks/utils/app';
 import { ActiveFilters } from '../../types/types';
 
 function FilterOptionsList(props: FilterOptionsListProps) {
