+++ src/components/CompareResults/FilterStatusChip.tsx
@@ -1,7 +1,7 @@
 import { Stack, Chip } from '@mui/material';
 
 import { RootState } from '../../common/store';
-import { useAppSelector } from '../../hooks/app';
+import { useAppSelector } from '../../hooks/utils/app';
 import { ActiveFilters } from '../../types/types';
 
 const FilterStatusChip = (props: FilterStatusChipProps) => {
