+++ src/components/Search/SearchViewInit.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { connect } from 'react-redux';
 
 import type { RootState } from '../../common/store';
-import { useAppDispatch } from '../../hooks/app';
+import { useAppDispatch } from '../../hooks/utils/app';
 import { fetchRecentRevisions } from '../../thunks/searchThunk';
 import type { Repository } from '../../types/state';
