+++ src/components/Shared/RevisionSearch.tsx
@@ -7,8 +7,8 @@ import Grid from '@mui/material/Grid';
 import { connect } from 'react-redux';
 
 import { RootState } from '../../common/store';
-import { useAppDispatch } from '../../hooks/app';
-import useSelectRevision from '../../hooks/useSelectRevision';
+import useSelectRevision from '../../hooks/Revisions/useSelectRevision';
+import { useAppDispatch } from '../../hooks/utils/app';
 import { clearCheckedRevisions } from '../../reducers/CheckedRevisions';
 import type { Revision } from '../../types/state';
 import AddRevisionButton from '../Search/AddRevisionButton';
