+++ src/hooks/Revisions/useCheckRevision.ts
@@ -1,15 +1,16 @@
 import { useSnackbar, VariantType } from 'notistack';
 
-import { setCheckedRevisions } from '../reducers/CheckedRevisions';
-import { Revision } from '../types/state';
-import { useAppDispatch, useAppSelector } from './app';
+import { RootState } from '../../common/store';
+import { setCheckedRevisions } from '../../reducers/CheckedRevisions';
+import { Revision } from '../../types/state';
+import { useAppDispatch, useAppSelector } from '../utils/app';
 
 const useCheckRevision = () => {
   const { enqueueSnackbar } = useSnackbar();
   const dispatch = useAppDispatch();
 
-  const checkedRevisions: Revision[] = useAppSelector(
-    (state) => state.checkedRevisions.revisions,
+  const checkedRevisions = useAppSelector(
+    (state: RootState) => state.checkedRevisions.revisions,
   );
 
   const handleToggle = (revision: Revision, maxRevisions: number) => {
