+++ src/hooks/utils/app.ts
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import type { TypedUseSelectorHook } from 'react-redux';
 
-import type { RootState, AppDispatch } from '../common/store';
+import type { RootState, AppDispatch } from '../../common/store';
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch: () => AppDispatch = useDispatch;
